% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd}
\alias{dmd}
\title{Dynamic Mode Decomposition (DMD)}
\usage{
dmd(x, y = NULL, svd.rank)
}
\arguments{
\item{x}{matrix of raw data snapshots}

\item{y}{matrix of output data snapshots (optional)}

\item{svd.rank;}{int rank of SVD}
}
\value{
dmd.obj; list containing lambda (vector of DMD eigenvalues), Phi (matrix of DMD modes) and b (vector of mode amplitudes)
}
\description{
Computes the dynamic mode decomposition of the data.
}
