% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd_reconst.R
\name{dmd_reconst}
\alias{dmd_reconst}
\title{DMD Reconstruction}
\usage{
dmd_reconst(dmd.obj)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}
}
\value{
X.recon; matrix, reconstructed approximation of X matrix
}
\description{
Reconstructs dynamic mode decomposition input data.
}
