% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\docType{data}
\name{.packageName}
\alias{.packageName}
\title{Dynamic Mode Decomposition (DMD)}
\format{
An object of class \code{character} of length 1.
}
\usage{
.packageName
}
\arguments{
\item{x}{matrix of raw data snapshots}

\item{y}{matrix of output data snapshots (optional)}

\item{svd.rank;}{int rank of SVD}
}
\value{
dmd.obj; list containing lambda (vector of DMD eigenvalues), Phi (matrix of DMD modes) and b (vector of mode amplitudes)
}
\description{
Computes the dynamic mode decomposition of the data.
}
\keyword{datasets}
