% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epanKernel}
\alias{epanKernel}
\title{Epanichnikov Kernel for One Dimension}
\arguments{
\item{u}{A numeric value representing the input.}
}
\value{
A numeric value representing the kernel evaluation.
}
\description{
Computes the Epanichnikov kernel value for a scalar input.
}
