% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epanKernelMulti}
\alias{epanKernelMulti}
\title{Multivariate Epanichnikov Kernel}
\arguments{
\item{u}{A numeric row vector representing the input.}
}
\value{
A numeric value representing the kernel evaluation.
}
\description{
Computes the product of Epanichnikov kernel values for a multivariate input.
}
