% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Gaussian Kernel}
\arguments{
\item{u}{A numeric row vector (p) representing the input data. Each element of this vector is used to calculate the corresponding Gaussian kernel value.}
}
\value{
A numeric value representing the result of the Gaussian kernel function applied to the input vector \code{u}. The value is calculated as the product of Gaussian functions applied element-wise to each component of \code{u}.
If the result is less than 0.01, it returns 0.01 to avoid extremely small values that might cause numerical instability in further computations.
}
\description{
This function computes the Gaussian kernel value for a given input vector. The kernel is applied element-wise to the input vector \code{u} and returns the product of the Gaussian functions of each element.
}
