% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eigs.discrete}
\alias{plot.eigs.discrete}
\title{Plot Eigenvalues (Discrete)}
\usage{
\method{plot}{eigs.discrete}(dmd.obj, save = FALSE)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots the DMD eigenvalues in discrete form as points on the complex plane.
}
