% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contour}
\alias{plot_contour}
\title{Contour Plot}
\usage{
plot_contour(X, complex.part = "real", save = FALSE)
}
\arguments{
\item{X;}{matrix, raw data matrix or dmd_reconst() output data matrix}

\item{complex.part;}{character,  desired mode for plotting, "real", "imaginary", "phase", or "magnitude"}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots data matrix using a filled contour plot.
}
