% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dynamics}
\alias{plot_dynamics}
\title{Plot Dynamics}
\usage{
plot_dynamics(dmd.obj, t, complex.part = "real", save = FALSE)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}

\item{t;}{vector, time values}

\item{complex.part;}{character,  desired mode for plotting, "real", "imaginary", "phase", or "magnitude"}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots underlying dynamics of all DMD modes.
}
