% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dynamics_num}
\alias{plot_dynamics_num}
\title{Plot Dynamics (number)}
\usage{
plot_dynamics_num(dmd.obj, t, mode.num, complex.part = "real", save = FALSE)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}

\item{t;}{vector, time values}

\item{mode.num;}{int, number of specified mode for plotting}

\item{complex.part;}{character,  desired mode for plotting, "real", "imaginary", "phase", or "magnitude"}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots the underlying dynamics of a specified DMD mode.
}
