% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modes}
\alias{plot_modes}
\title{Plot Modes}
\usage{
plot_modes(dmd.obj, x, complex.part = "real", save = FALSE)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}

\item{x;}{vector, x values}

\item{complex.part;}{character,  desired mode for plotting, "real", "imaginary", "phase", or "magnitude"}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots all DMD modes.
}
