% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_modes_num}
\alias{plot_modes_num}
\title{Plot Mode (number)}
\usage{
plot_modes_num(dmd.obj, x, mode.num, complex.part = "real", save = FALSE)
}
\arguments{
\item{dmd.obj;}{list, output from dmd() function}

\item{x;}{vector, x values}

\item{mode.num;}{int, number of specified mode for plotting}

\item{complex.part;}{character,  desired mode for plotting, "real", "imaginary", "phase", or "magnitude"}

\item{save;}{boolean, saves plot if TRUE, prints plot if FALSE}
}
\value{
p; plot
}
\description{
Plots a specified DMD mode.
}
