% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trCor}
\alias{trCor}
\title{Calculate Trace Correlation Between Subspaces}
\usage{
trCor(beta, beta_hat)
}
\arguments{
\item{beta}{A numeric matrix (\( m \times n \)), defining the first subspace.}

\item{beta_hat}{A numeric matrix (\( m \times p \)), defining the second subspace.}
}
\value{
A numeric scalar representing the trace correlation:
\deqn{\text{trCor} = \frac{\text{tr}(P_{\text{beta\_hat}} P_{\text{beta}})}{n}}
where \( P_{\text{beta}} \) and \( P_{\text{beta\\emph{hat}} \) are the projection matrices for \( \beta \) and \( \beta}{\text{hat}} \), respectively, and \( n \) is the number of columns in \( \beta \).
}
\description{
Computes the trace correlation between two subspaces defined by \( \beta \) and \( \beta_{\text{hat}} \), based on their projection matrices.
}
\details{
Trace correlation measures the similarity between two subspaces by comparing their projection matrices. It ranges from 0 to 1, where higher values indicate greater similarity. The function internally uses the \code{projection} function to compute the required projection matrices.
}
\examples{
# Example in R
beta <- matrix(rnorm(20), nrow = 5)
beta_hat <- matrix(rnorm(15), nrow = 5)
trace_correlation <- trCor(beta, beta_hat)
print(trace_correlation)

}
