% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecInv}
\alias{vecInv}
\title{Inverse Operation of vecLower}
\arguments{
\item{v}{A numeric vector representing the lower triangular block.}

\item{p}{Number of rows in the reconstructed matrix.}

\item{d}{Number of columns in the triangular block.}
}
\value{
A numeric matrix reconstructed from the input vector.
}
\description{
Reconstructs a matrix from its vectorized lower triangular block.
}
