% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vecLower}
\alias{vecLower}
\title{Extract Lower Triangular Block of a Matrix}
\arguments{
\item{m}{A numeric matrix input.}

\item{p}{Number of rows in the matrix.}

\item{d}{Number of columns in the triangular block.}
}
\value{
A numeric vector representing the vectorized lower triangular block.
}
\description{
Extracts and vectorizes the lower triangular portion of a matrix.
}
