% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimateScore}
\alias{estimateScore}
\title{Estimate Scores and Parameters}
\usage{
estimateScore(X, alpha, Y, h1, h2, h3)
}
\arguments{
\item{X}{A numeric matrix (n x p) of predictors.}

\item{alpha}{A numeric matrix (p x d) of transformation coefficients.}

\item{Y}{A numeric vector (n) of responses.}

\item{h1}{Bandwidth parameter for regression kernel.}

\item{h2}{Bandwidth parameter for weight estimation kernel.}

\item{h3}{Bandwidth parameter for covariate adjustment kernel.}

\item{weights}{A numeric vector (n) of observation weights.}

\item{isCentral}{A boolean indicating whether to compute centralized scores.}
}
\value{
A list containing:
\itemize{
\item \code{mHat}: Vector of estimated intercepts.
\item \code{m1Hat}: Matrix of estimated slope coefficients.
\item \code{eWHat}: Vector of weighted averages of responses.
\item \code{exWHat}: Matrix of weighted averages of predictors.
\item \code{scoreHat}: Matrix of estimated scores.
\item \code{eScore}: Vector of mean scores.
\item \code{H}: Matrix of estimated Hessians.
}
}
\description{
Implements the main algorithm for estimating scores and parameters.
}
